refactor: module 이름 변경
