+++ app/src/main/java/woowacourse/shopping/ShoppingApp.kt
@@ -1,12 +1,12 @@
 package woowacourse.shopping
 
 import android.app.Application
-import woowacourse.shopping.di.AppModule
+import woowacourse.shopping.di.RepositoryModule
 import woowacourse.shopping.di.DIContainer
 
 class ShoppingApp : Application() {
     override fun onCreate() {
         super.onCreate()
-        DIContainer.loadModule(AppModule())
+        DIContainer.loadModule(RepositoryModule())
     }
 }
